.history/requirements_20250814114548.txt
gTTS==2.5.4
moviepy==1.0.3
numpy==2.3.2
openai==1.99.9
pillow==11.3.0
praw==7.8.1
python-dotenv==1.1.1
requests==2.32.4
